+++ src/pipp_ser_write.h
@@ -144,7 +144,7 @@ static_assert (sizeof(s_ser_header) == 7 * 4 + 3 * 40 + 2 * 8, "Unexpected size 
             uint8_t *p_read = (uint8_t *)&data;
             p_read += sizeof(data) - 1;
             uint8_t *p_write = (uint8_t *)&ret;
-            for (int x = 0; x < sizeof(data); x++) {
+            for (size_t x = 0; x < sizeof(data); x++) {
                 *p_write++ = *p_read--;
             }
